refactor(models): move packed-batch layout helpers next to their only consumer
pack_padded_batch/unpack_to_padded serve BaseModel's input preparation, not the
FSDP strategy layer; molt/trainer/fsdp/packing.py becomes molt/models/packing.py.
